+++ preproc/write_schism_ic.m
@@ -1,5 +1,5 @@
 function write_schism_ic(Mobj, prefix_name, varData)
-% Write the *ic files for SCHISM
+% Write the *ic file for SCHISM
 % 
 %% Syntax
 % write_schism_ic(Mobj, prefix_name, varData)
@@ -11,14 +11,18 @@ function write_schism_ic(Mobj, prefix_name, varData)
 %% Example
 % ssh = rand(Mobj.nNodes ,1);
 % write_schism_ic(Mobj, 'elev', ssh)
-%
 % write_schism_ic(Mobj, 'salt', 30)
 %
 %% Input Arguments
-% Mobj --- the mesh object
-% prefix_name --- prefix name for the ic file.
-% varData --- variable vector or scalar. If the input is a scalar, it will
-% be expanded to a Mobj.nNodes*1 vector automatically.
+% Mobj - mesh object; datastruct
+%       the datastruct used to store mesh info.
+% prefix_name - prefix name; char
+%       the file prefix name. e.g. prefix_name = 'drag', or 'drag.gr3'. 
+%       the file extension will be omitted automatically.
+% varData - input data; numeric
+%       the input data to be written. If '' is a scalar, it will
+%       be expanded to a vector automatically , and the data at all nodes
+%       are the same.in this case.
 %
 %% Output Arguments
 % None
@@ -28,19 +32,20 @@ function write_schism_ic(Mobj, prefix_name, varData)
 %
 %% Author Info
 % Created by Wenfan Wu, Virginia Institute of Marine Science in 2021. 
-% Last Updated on 6 Nov. 2024. 
+% Last Updated on 5 Apr 2025. 
 % Email: wwu@vims.edu
 % 
 % See also:write_schism_gr3 and write_schism_hotstart
 
 %% Parse inputs
-if isscalar(varData)
-    varData = varData*ones(1, Mobj.nNodes);
-end
-head_line = datestr(Mobj.time(1), 'mmm/dd/yyyy'); %#ok<*DATST>
 prefix_name = regexprep(prefix_name, '\.ic$', '');
+filepath = fullfile(Mobj.aimpath, [prefix_name, '.ic']);
+head_line = datestr(now, 'mmm/dd/yyyy HH:MM:SS'); %#ok<TNOW1,DATST>
 
 %% Check
+if isscalar(varData); varData = varData*ones(Mobj.nNodes, 1); end
+if any(isnan(varData(:))); error('NaNs were found in the input data!'); end
+
 switch prefix_name
     case 'temp'
         if min(varData)<-2 || max(varData)>40
@@ -58,26 +63,19 @@ switch prefix_name
         end
 end
 %% Begin to write
-fileName = fullfile(Mobj.aimpath, [prefix_name, '.ic']);
-fid = fopen(fileName,'wt');
+fid = fopen(filepath,'w');
 fprintf(fid, [head_line, '\n']);
-fprintf(fid, [num2str(Mobj.nElems),' ',num2str(Mobj.nNodes), '\n']);
+fprintf(fid,'%d %d\n', Mobj.nElems, Mobj.nNodes);
 
-% Node Info
-node_part = [(1:Mobj.nNodes)', Mobj.lon(:), Mobj.lat(:), varData(:)]';
-node_fmt = repmat('%d   %14.6f   %14.6f   %13.7e\n', 1, size(node_part,2));
-fprintf(fid, node_fmt, node_part(:));
+% Node and Elem
+fprintf(fid, '%d   %14.6f   %14.6f    %13.7e\n', [(1:Mobj.nNodes)', Mobj.lon(:), Mobj.lat(:), varData(:)]');
+fprintf(fid, '%d %d %d %d %d %d\n', [(1:Mobj.nElems)', Mobj.i34(:), Mobj.tri]');  % include NaN values
+fclose(fid);
 
-% Elem Info
-elem_part = [(1:Mobj.nElems)', Mobj.i34(:).*ones(Mobj.nElems,1), Mobj.tri]';
-elem_part_1d = elem_part(:);
-elem_part_1d(isnan(elem_part_1d)) = [];
-
-elem_fmt = repmat('%d %d %d %d %d %d\n', Mobj.nElems, 1);
-elem_fmt(Mobj.i34==3, :) = repmat('%d %d %d %d %d   \n', sum(Mobj.i34==3), 1);
-elem_fmt = elem_fmt'; elem_fmt_1d = elem_fmt(:);
-fprintf(fid, elem_fmt_1d, elem_part_1d);
-
+% Remove NaNs
+fileText = fileread(filepath);
+fid = fopen(filepath, 'w');
+fwrite(fid, strrep(fileText, 'NaN', '')); % case-sensitive
 fclose(fid);
 
 disp([prefix_name, '.ic has been created successfully!'])